refactor: Player, Players, ExecutionResults 를 engine으로 이동
