+++ src/main/java/ladder/domain/engine/LadderResults.java
@@ -1,7 +1,5 @@
 package ladder.domain.engine;
 
-import ladder.domain.nextstep.Player;
-
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
