+++ src/main/java/ladder/domain/engine/Player.java
@@ -1,4 +1,4 @@
-package ladder.domain.nextstep;
+package ladder.domain.engine;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
